bitwise operators - lists.h to main.h

+++ 0x14-bit_manipulation/4-clear_bit.c
@@ -1,4 +1,4 @@
-#include "lists.h"
+#include "main.h"
 
 /**
  * clear_bit - sets the value of a bit to 0 at a given index
@@ -13,7 +13,7 @@ int clear_bit(unsigned long int *n, unsigned int index)
 
 	if (n == NULL)
 		return (-1);
-	if ((1 << index) > *n)
+	if ((1 << index) > (int)*n)
 		return (-1);
 	if ((*n >> index) % 2 == 1)
 	{
